[clang][Index] Split USRGeneration out of clangIndex into clangUnifiedSymbolResolution
Previously, USRGeneration was implemented by clangIndex.
However, that poses too broad library layering constraints on the ever
growing set of users of a tiny component of it, USRGeneration.

This PR splits that into a small library, called: clangUnifiedSymbolResolution

Anyone needing USRGeneration could simply link against this without
pulling in everything from clangIndex.

---

Importantly, clangIndex linked against clangFrontend to define its
FrontendAction, and use some ASTUnit APIs. Some users may want to use
USRGeneration but NOT depend on clangFrontend. This new
clangUnifiedSymbolResolution library would solve such circular
dependencies.

PS: There were quite a few cases where libraries could just link
against clangUnifiedSymbolResolution without linking to clangIndex.
I've simplified those cases in this PR, to keep link deps minimal.
