+++ clang/include/clang/UnifiedSymbolResolution/USRGeneration.h
