+++ clang/lib/Analysis/Scalable/ASTEntityMapping.cpp
@@ -13,7 +13,7 @@
 #include "clang/Analysis/Scalable/ASTEntityMapping.h"
 #include "clang/AST/Decl.h"
 #include "clang/Analysis/Scalable/Model/BuildNamespace.h"
-#include "clang/Index/USRGeneration.h"
+#include "clang/UnifiedSymbolResolution/USRGeneration.h"
 #include "llvm/ADT/SmallString.h"
 
 namespace clang::ssaf {
