+++ clang/lib/CMakeLists.txt
@@ -30,6 +30,7 @@ if(CLANG_INCLUDE_TESTS)
 endif()
 add_subdirectory(Interpreter)
 add_subdirectory(Support)
+add_subdirectory(UnifiedSymbolResolution)
 
 if(CLANG_ENABLE_CIR)
   add_subdirectory(CIR)
