+++ clang/lib/Index/CMakeLists.txt
@@ -12,7 +12,6 @@ add_clang_library(clangIndex
   IndexingContext.cpp
   IndexSymbol.cpp
   IndexTypeSourceInfo.cpp
-  USRGeneration.cpp
 
   ADDITIONAL_HEADERS
   IndexingContext.h
@@ -26,6 +25,7 @@ add_clang_library(clangIndex
   clangSema
   clangSerialization
   clangToolingCore
+  clangUnifiedSymbolResolution
 
   DEPENDS
   omp_gen
