+++ clang/lib/StaticAnalyzer/Core/CMakeLists.txt
@@ -61,10 +61,10 @@ add_clang_library(clangStaticAnalyzerCore
   clangBasic
   clangCrossTU
   clangFrontend
-  clangIndex
   clangLex
   clangRewrite
   clangToolingCore
+  clangUnifiedSymbolResolution
 
   DEPENDS
   omp_gen
