+++ clang/lib/UnifiedSymbolResolution/CMakeLists.txt
@@ -1,0 +1,15 @@
+set(LLVM_LINK_COMPONENTS
+  Support
+  )
+
+add_clang_library(clangUnifiedSymbolResolution
+  USRGeneration.cpp
+
+  DEPENDS
+  omp_gen
+
+  LINK_LIBS
+  clangAST
+  clangBasic
+  clangLex
+  )
