+++ clang/tools/c-index-test/CMakeLists.txt
@@ -31,6 +31,7 @@ else()
     clangFrontend
     clangIndex
     clangSerialization
+    clangUnifiedSymbolResolution
   )
 endif()
